producer: return [in out], take :duplex? and records w/ :response-ch
Producer now returns a vector of two channels [in out].
Producer now accepts a :duplex?` boolean config entry which makes it
yield a duplex channel instead of a vector of two channels.
Producer record maps can now have an optional :response-ch entry
with a channel which will convey the send result as a map containing
a `:type` key where `:type` may be:, either:
  - `:exception`: the exception in entry `:exception` was raised
  - `:record-metadata`: a producer record map in the form:
    {:type      :record-metadata
     :topic     \"topic-name\"
     :partition 0 ;; nil if unknown
     :offset    1234567890
     :timestamp 9876543210}
When a response channel is provided, results are not sent to the
out/duplex channel, only to the response channel in a blocking manner,
so make sure the response channel is consumed.
The response channel will be closed after the operation.